Corrects typos. Includes version info.
The presented poster had images 6c and 6d switched in the caption. This
typo has been corrected.

The subscript ``r'' on eta in the y label of Figure 7 was a typo. This
has been removed so that it is now simply eta_1.

Installation of the style sheet was not meant to be a default. This typo
has been commented out so that it only runs when the user desires.

Code folding was apparently changed. This should not effect content.

Version information is not cleared so that others can see what versions
were used.